+++ src/components/options/BlocklistTab.tsx
@@ -1,36 +1,20 @@
 import React, { useState, useCallback } from 'react';
-import {
-  Plus,
-  Trash2,
-  Shield,
-  Clock,
-  ChevronDown,
-  ChevronUp,
-  Bell,
-  Lock,
-  Youtube,
-  PlaySquare,
-  ThumbsUp,
-  MessageSquare,
-  LayoutPanelLeft,
-  Home,
-  Info
-} from 'lucide-react';
+import { Plus, Lock } from 'lucide-react';
 
-import { Button, Card, Input, Toggle, Select } from '~/components/ui';
+import { Button, Card, Input } from '~/components/ui';
 import { PasswordModal } from '~/components/options/modals';
-import { TimeLimitBadge } from '~/components/features';
 import { getMessage } from '~/lib/i18n';
-import { TIME_LIMIT_CONFIG } from '~/constants/limits';
+import {
+  NotificationSettingsSection,
+  YouTubeSection,
+  DomainListItem
+} from '~/components/options/blocklist';
 import type {
   AppSettings,
   SiteBlockCount,
-  BlockItem,
   TimeLimit,
-  TimeLimitType,
   TimeLimitUsage,
   NotificationSettings,
-  NotificationMinutes,
   YouTubeSettings
 } from '~/types/storage';
 
@@ -48,425 +32,6 @@ interface BlocklistTabProps {
   timeLimitUsage?: Record<string, TimeLimitUsage>;
   youtube: YouTubeSettings;
   onYouTubeChange: (youtube: YouTubeSettings) => void;
-}
-
-type LimitTypeOption = 'always' | 'daily' | 'hourly';
-
-interface TimeLimitEditorProps {
-  item: BlockItem;
-  onUpdate: (timeLimit: TimeLimit | null) => void;
-  usage?: TimeLimitUsage;
-}
-
-function TimeLimitEditor({ item, onUpdate, usage }: TimeLimitEditorProps) {
-  const [isExpanded, setIsExpanded] = useState(false);
-
-  const currentType: LimitTypeOption = item.timeLimit
-    ? item.timeLimit.type
-    : 'always';
-  const currentMinutes = item.timeLimit
-    ? Math.floor(item.timeLimit.limitSeconds / 60)
-    : currentType === 'daily'
-      ? TIME_LIMIT_CONFIG.DEFAULT_DAILY_LIMIT / 60
-      : TIME_LIMIT_CONFIG.DEFAULT_HOURLY_LIMIT / 60;
-
-  const [selectedType, setSelectedType] = useState<LimitTypeOption>(currentType);
-  const [minutes, setMinutes] = useState(currentMinutes);
-
-  const handleTypeChange = useCallback(
-    (newType: LimitTypeOption) => {
-      setSelectedType(newType);
-
-      if (newType === 'always') {
-        onUpdate(null);
-      } else {
-        const defaultMinutes =
-          newType === 'daily'
-            ? TIME_LIMIT_CONFIG.DEFAULT_DAILY_LIMIT / 60
-            : TIME_LIMIT_CONFIG.DEFAULT_HOURLY_LIMIT / 60;
-        setMinutes(defaultMinutes);
-        onUpdate({
-          type: newType as TimeLimitType,
-          limitSeconds: defaultMinutes * 60
-        });
-      }
-    },
-    [onUpdate]
-  );
-
-  const handleMinutesChange = useCallback(
-    (value: string) => {
-      const newMinutes = parseInt(value, 10);
-      if (isNaN(newMinutes) || newMinutes < 1) return;
-
-      const clampedMinutes = Math.min(
-        Math.max(newMinutes, TIME_LIMIT_CONFIG.MIN_LIMIT_SECONDS / 60),
-        TIME_LIMIT_CONFIG.MAX_LIMIT_SECONDS / 60
-      );
-      setMinutes(clampedMinutes);
-
-      if (selectedType !== 'always') {
-        onUpdate({
-          type: selectedType as TimeLimitType,
-          limitSeconds: clampedMinutes * 60
-        });
-      }
-    },
-    [selectedType, onUpdate]
-  );
-
-  const typeOptions = [
-    { value: 'always', label: getMessage('alwaysBlocked') },
-    { value: 'daily', label: getMessage('dailyLimit') },
-    { value: 'hourly', label: getMessage('hourlyLimit') }
-  ];
-
-  // Calculate remaining time for display
-  const remainingSeconds =
-    item.timeLimit && usage
-      ? item.timeLimit.limitSeconds -
-        (item.timeLimit.type === 'daily'
-          ? usage.dailyUsedSeconds
-          : usage.hourlyUsedSeconds)
-      : null;
-
-  return (
-    <div className="mt-2">
-      <button
-        onClick={() => setIsExpanded(!isExpanded)}
-        className="flex items-center gap-1 text-xs text-gray-500 hover:text-gray-700"
-      >
-        <Clock className="w-3 h-3" />
-        {item.timeLimit ? (
-          <span>
-            {getMessage('limitMinutes', minutes.toString())}
-            {item.timeLimit.type === 'daily'
-              ? getMessage('perDay')
-              : getMessage('perHour')}
-          </span>
-        ) : (
-          <span>{getMessage('alwaysBlocked')}</span>
-        )}
-        {isExpanded ? (
-          <ChevronUp className="w-3 h-3" />
-        ) : (
-          <ChevronDown className="w-3 h-3" />
-        )}
-      </button>
-
-      {isExpanded && (
-        <div className="mt-2 p-3 bg-gray-50 rounded-lg space-y-3">
-          <div>
-            <label className="block text-xs font-medium text-gray-600 mb-1">
-              {getMessage('timeLimitType')}
-            </label>
-            <Select
-              value={selectedType}
-              onChange={(value) => handleTypeChange(value as LimitTypeOption)}
-              options={typeOptions}
-            />
-          </div>
-
-          {selectedType !== 'always' && (
-            <div>
-              <label className="block text-xs font-medium text-gray-600 mb-1">
-                {getMessage('timeLimitDuration')}
-              </label>
-              <div className="flex items-center gap-2">
-                <Input
-                  type="number"
-                  value={minutes.toString()}
-                  onChange={handleMinutesChange}
-                  min={1}
-                  max={TIME_LIMIT_CONFIG.MAX_LIMIT_SECONDS / 60}
-                  className="w-24"
-                />
-                <span className="text-sm text-gray-500">
-                  {getMessage('minutes')}
-                </span>
-              </div>
-              <p className="text-xs text-gray-400 mt-1">
-                {selectedType === 'daily'
-                  ? getMessage('resetDaily')
-                  : getMessage('resetHourly')}
-              </p>
-            </div>
-          )}
-        </div>
-      )}
-
-      {/* Show remaining time badge if time limit is set and item is enabled */}
-      {item.enabled && item.timeLimit && remainingSeconds !== null && (
-        <div className="mt-2">
-          <TimeLimitBadge
-            remainingSeconds={Math.max(0, remainingSeconds)}
-            limitSeconds={item.timeLimit.limitSeconds}
-            limitType={item.timeLimit.type}
-            compact
-          />
-        </div>
-      )}
-    </div>
-  );
-}
-
-// Notification settings section component
-interface NotificationSettingsProps {
-  notifications: NotificationSettings | undefined;
-  onUpdate: (notifications: NotificationSettings) => void;
-  hasTimeLimitSites: boolean;
-}
-
-function NotificationSettingsSection({
-  notifications,
-  onUpdate,
-  hasTimeLimitSites
-}: NotificationSettingsProps) {
-  const timeLimitEnabled = notifications?.timeLimitEnabled ?? true;
-  const timeLimitMinutes = notifications?.timeLimitMinutes ?? 5;
-
-  const minuteOptions: Array<{ value: string; label: string }> = [
-    { value: '1', label: getMessage('notificationMinutesOption', '1') },
-    { value: '3', label: getMessage('notificationMinutesOption', '3') },
-    { value: '5', label: getMessage('notificationMinutesOption', '5') },
-    { value: '10', label: getMessage('notificationMinutesOption', '10') }
-  ];
-
-  const handleToggle = useCallback(
-    (enabled: boolean) => {
-      onUpdate({
-        timeLimitEnabled: enabled,
-        timeLimitMinutes
-      });
-    },
-    [onUpdate, timeLimitMinutes]
-  );
-
-  const handleMinutesChange = useCallback(
-    (value: string) => {
-      const minutes = parseInt(value, 10) as NotificationMinutes;
-      onUpdate({
-        timeLimitEnabled,
-        timeLimitMinutes: minutes
-      });
-    },
-    [onUpdate, timeLimitEnabled]
-  );
-
-  // Only show if there are sites with time limits
-  if (!hasTimeLimitSites) {
-    return null;
-  }
-
-  return (
-    <Card>
-      <div className="flex items-center gap-3 mb-4">
-        <Bell className="w-5 h-5 text-blue-500" />
-        <h2 className="text-lg font-semibold text-gray-900">
-          {getMessage('notificationSettings')}
-        </h2>
-      </div>
-
-      <div className="space-y-4">
-        <div className="flex items-center justify-between">
-          <div>
-            <p className="font-medium text-gray-900">
-              {getMessage('notificationTimeLimitEnabled')}
-            </p>
-            <p className="text-sm text-gray-500">
-              {getMessage('notificationTimeLimitEnabledDescription')}
-            </p>
-          </div>
-          <Toggle
-            checked={timeLimitEnabled}
-            onChange={handleToggle}
-          />
-        </div>
-
-        {timeLimitEnabled && (
-          <div>
-            <label className="block text-sm font-medium text-gray-600 mb-2">
-              {getMessage('notificationTimeLimitMinutes')}
-            </label>
-            <Select
-              value={timeLimitMinutes.toString()}
-              onChange={handleMinutesChange}
-              options={minuteOptions}
-              className="w-48"
-            />
-          </div>
-        )}
-      </div>
-    </Card>
-  );
-}
-
-// YouTube feature toggle item
-interface YouTubeFeatureToggleProps {
-  icon: React.ReactNode;
-  title: string;
-  description: string;
-  checked: boolean;
-  onChange: (checked: boolean) => void;
-  disabled?: boolean;
-}
-
-function YouTubeFeatureToggle({
-  icon,
-  title,
-  description,
-  checked,
-  onChange,
-  disabled = false
-}: YouTubeFeatureToggleProps) {
-  return (
-    <div
-      className={`flex items-start gap-3 p-3 rounded-lg transition-colors ${
-        disabled
-          ? 'bg-gray-50 opacity-60'
-          : checked
-            ? 'bg-red-50'
-            : 'bg-gray-50 hover:bg-gray-100'
-      }`}
-    >
-      <div
-        className={`p-1.5 rounded-lg ${checked && !disabled ? 'bg-red-100 text-red-600' : 'bg-gray-200 text-gray-500'}`}
-      >
-        {icon}
-      </div>
-      <div className="flex-1 min-w-0">
-        <div className="flex items-center justify-between gap-2">
-          <h4
-            className={`text-sm font-medium ${disabled ? 'text-gray-400' : 'text-gray-900'}`}
-          >
-            {title}
-          </h4>
-          <Toggle
-            checked={checked}
-            onChange={onChange}
-            disabled={disabled}
-            size="sm"
-          />
-        </div>
-        <p
-          className={`text-xs mt-0.5 ${disabled ? 'text-gray-400' : 'text-gray-500'}`}
-        >
-          {description}
-        </p>
-      </div>
-    </div>
-  );
-}
-
-// YouTube settings section
-interface YouTubeSectionProps {
-  youtube: YouTubeSettings;
-  onYouTubeChange: (youtube: YouTubeSettings) => void;
-}
-
-function YouTubeSection({ youtube, onYouTubeChange }: YouTubeSectionProps) {
-  const isEnabled = youtube?.enabled ?? false;
-
-  const handleToggle = useCallback(
-    (key: keyof YouTubeSettings) => (checked: boolean) => {
-      onYouTubeChange({ ...youtube, [key]: checked });
-    },
-    [youtube, onYouTubeChange]
-  );
-
-  const features = [
-    {
-      key: 'hideShorts' as const,
-      icon: <PlaySquare className="w-4 h-4" />,
-      title: getMessage('youtubeHideShorts'),
-      description: getMessage('youtubeHideShortsDescription')
-    },
-    {
-      key: 'hideRecommendations' as const,
-      icon: <ThumbsUp className="w-4 h-4" />,
-      title: getMessage('youtubeHideRecommendations'),
-      description: getMessage('youtubeHideRecommendationsDescription')
-    },
-    {
-      key: 'hideComments' as const,
-      icon: <MessageSquare className="w-4 h-4" />,
-      title: getMessage('youtubeHideComments'),
-      description: getMessage('youtubeHideCommentsDescription')
-    },
-    {
-      key: 'hideSidebar' as const,
-      icon: <LayoutPanelLeft className="w-4 h-4" />,
-      title: getMessage('youtubeHideSidebar'),
-      description: getMessage('youtubeHideSidebarDescription')
-    },
-    {
-      key: 'hideHomeFeed' as const,
-      icon: <Home className="w-4 h-4" />,
-      title: getMessage('youtubeHideHomeFeed'),
-      description: getMessage('youtubeHideHomeFeedDescription')
-    }
-  ];
-
-  return (
-    <Card>
-      <div className="flex items-start gap-3 mb-4">
-        <div className="p-2 bg-red-100 rounded-lg">
-          <Youtube className="w-5 h-5 text-red-600" />
-        </div>
-        <div className="flex-1">
-          <h2 className="text-lg font-semibold text-gray-900">
-            {getMessage('youtubeBlockingTitle')}
-          </h2>
-          <p className="text-sm text-gray-500 mt-0.5">
-            {getMessage('youtubeBlockingDescription')}
-          </p>
-        </div>
-      </div>
-
-      {/* Master Toggle */}
-      <div className="p-3 bg-gray-50 rounded-lg mb-4">
-        <div className="flex items-center justify-between">
-          <div>
-            <h3 className="font-medium text-gray-900">
-              {getMessage('youtubeEnabled')}
-            </h3>
-            <p className="text-xs text-gray-500">
-              {getMessage('youtubeEnabledDescription')}
-            </p>
-          </div>
-          <Toggle
-            checked={isEnabled}
-            onChange={handleToggle('enabled')}
-            size="lg"
-          />
-        </div>
-      </div>
-
-      {/* Feature Toggles */}
-      {!isEnabled && (
-        <div className="mb-3 p-2 bg-amber-50 border border-amber-200 rounded-lg flex items-center gap-2">
-          <Info className="w-4 h-4 text-amber-600 flex-shrink-0" />
-          <p className="text-xs text-amber-700">
-            {getMessage('youtubeDisabledNote')}
-          </p>
-        </div>
-      )}
-
-      <div className="space-y-2">
-        {features.map((feature) => (
-          <YouTubeFeatureToggle
-            key={feature.key}
-            icon={feature.icon}
-            title={feature.title}
-            description={feature.description}
-            checked={youtube?.[feature.key] ?? false}
-            onChange={handleToggle(feature.key)}
-            disabled={!isEnabled}
-          />
-        ))}
-      </div>
-    </Card>
-  );
 }
 
 export function BlocklistTab({
@@ -540,8 +105,9 @@ export function BlocklistTab({
 
   // Check if any sites have time limits configured
   const hasTimeLimitSites =
-    settings?.blockList.some((item) => item.timeLimit !== null && item.timeLimit !== undefined) ??
-    false;
+    settings?.blockList.some(
+      (item) => item.timeLimit !== null && item.timeLimit !== undefined
+    ) ?? false;
 
   return (
     <div className="space-y-6">
@@ -595,58 +161,15 @@ export function BlocklistTab({
                 timeLimitUsage[domainKey] ?? timeLimitUsage[item.domain];
 
               return (
-                <div key={item.id} className="py-3">
-                  <div className="flex items-center justify-between">
-                    <div className="flex items-center gap-3">
-                      <Toggle
-                        checked={item.enabled}
-                        onChange={(checked) => handleToggleClick(item.id, checked)}
-                        size="sm"
-                      />
-                      <div>
-                        <p
-                          className={`font-medium ${item.enabled ? 'text-gray-900' : 'text-gray-400'}`}
-                        >
-                          {item.isWildcard && (
-                            <span
-                              className={
-                                item.enabled ? 'text-blue-600' : 'text-blue-300'
-                              }
-                            >
-                              *.
-                            </span>
-                          )}
-                          {item.domain}
-                        </p>
-                        <p className="text-xs text-gray-500">
-                          Added {new Date(item.createdAt).toLocaleDateString()}
-                        </p>
-                      </div>
-                      {blockCount > 0 && (
-                        <span className="inline-flex items-center gap-1 px-2 py-0.5 bg-red-100 text-red-700 text-xs rounded-full">
-                          <Shield className="w-3 h-3" />
-                          {getMessage('blockedTimesShort', blockCount.toString())}
-                        </span>
-                      )}
-                    </div>
-                    <Button
-                      variant="ghost"
-                      size="sm"
-                      onClick={() => handleRemoveClick(item.id)}
-                    >
-                      <Trash2 className="w-4 h-4 text-red-500" />
-                    </Button>
-                  </div>
-
-                  {/* Time Limit Editor */}
-                  <div className="ml-11">
-                    <TimeLimitEditor
-                      item={item}
-                      onUpdate={(timeLimit) => onUpdateTimeLimit(item.id, timeLimit)}
-                      usage={usage}
-                    />
-                  </div>
-                </div>
+                <DomainListItem
+                  key={item.id}
+                  item={item}
+                  blockCount={blockCount}
+                  usage={usage}
+                  onToggle={handleToggleClick}
+                  onRemove={handleRemoveClick}
+                  onUpdateTimeLimit={onUpdateTimeLimit}
+                />
               );
             })}
           </div>
